apps/dumili/api/paddleocr/requirements.txt
paddleocr>=3.4.0
paddlepaddle>=3.2.0
paddlex>=3.3.0
